+++ tests/e2e/heal_receipt_api_test.go
@@ -127,7 +127,7 @@ func TestE2E_HealReceiptAPI_3Node(t *testing.T) {
 			"--scrub-interval", "0",
 			"--lifecycle-interval", "0",
 			"--no-encryption",
-			"--ec=false",
+			"--cluster-ec=false",
 		)
 		cmd.Stdout = os.Stdout
 		cmd.Stderr = os.Stderr